requirements.txt
redis==2.9.1
qr==0.6.0
mongokit==0.9.1.1
nose==1.3.0
requests==2.2.1
rfc3987==1.3.4
elasticsearch==1.0.0
facepy==0.9.0
python-dateutil==2.2
google-api-python-client==1.2
oauth2client==1.2